dynamically showing incoming / outgoing notifications
