+++ store/index.js
@@ -6,13 +6,15 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import user from './user';
 import runs from './runs';
 import singleRun from './singleRun';
-import notifications from './notifications';
+import outgoing from './outgoing';
+import incoming from './incoming';
 
 const rootReducer = combineReducers({
   user: user,
   runs: runs,
   singleRun: singleRun,
-  notifications: notifications,
+  outgoing: outgoing,
+  incoming: incoming,
 });
 const middleware = composeWithDevTools(
   // applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
